requirements.txt
opencv-python
ultralytics
pillow
tkintertable
numpy
pandas
matplotlib
seaborn
